bug fix with entity id's, also added a bool check to entities to see if they were destroyed

+++ SerenityLib/Engine/src/ecs/ECS.h
@@ -7,7 +7,6 @@ namespace Engine::priv {
     struct InternalScenePublicInterface;
 }
 
-//#include <core/engine/utils/Utils.h>
 #include <ecs/ECSRegistry.h>
 #include <ecs/ECSEntityPool.h>
 #include <ecs/ECSSystem.h>
@@ -24,28 +23,7 @@ namespace Engine::priv {
             std::vector<Engine::priv::sparse_set_base*> m_ComponentPools;
             std::vector<ECSSystem<ENTITY>*>             m_Systems;
             std::mutex                                  m_Mutex;
-            /*
-            //builds a component pool and system for the component type if it is not built already.
-            template<typename COMPONENT> 
-            void buildPool(std::uint32_t type_slot) {
-                if (type_slot >= m_ComponentPools.size()) {
-                    m_ComponentPools.resize(type_slot + 1, nullptr);
-                }
-                if (m_ComponentPools[type_slot]) {
-                    return;
-                }
-                if (type_slot >= m_Systems.size()) {
-                    m_Systems.resize(type_slot + 1, nullptr);
-                }
-                if (!m_ComponentPools[type_slot]) {
-                    m_ComponentPools[type_slot] = NEW ECSComponentPool<ENTITY, COMPONENT>();
-                }
-                if (!m_Systems[type_slot]) {
-                    ECSSystemCI _ci;
-                    m_Systems[type_slot] = NEW ECSSystem<ENTITY, COMPONENT>(_ci, *this);
-                }
-            }
-            */
+
         public:
             ECS(/*const SceneOptions& options*/) {
             }
@@ -57,6 +35,10 @@ namespace Engine::priv {
             ECS& operator=(const ECS&)           = delete;
             ECS(ECS&& other) noexcept            = delete;
             ECS& operator=(ECS&& other) noexcept = delete;
+
+            const ECSEntityPool<ENTITY>& getEntityPool() const {
+                return m_EntityPool;
+            }
 
 
             //"event handlers"
@@ -105,18 +87,19 @@ namespace Engine::priv {
             }
             //destroy flagged entities & their components, if any
             void postUpdate(Scene& scene, const float dt) {
-                for (auto& entity : m_DestroyedEntities) {
-                    EntityDataRequest request(entity);
-                    auto id = request.ID;
-                    m_EntityPool.destroyFlaggedEntity(id);
-                    for (size_t i = 0; i < m_ComponentPools.size(); ++i) {
-                        m_ComponentPools[i]->remove(id);
-                        m_Systems[i]->onComponentRemovedFromEntity(entity);
+                if (m_DestroyedEntities.size() > 0) {
+                    for (const auto entity : m_DestroyedEntities) {
+                        auto id = entity.id();
+                        m_EntityPool.destroyFlaggedEntity(id);
+                        for (size_t i = 0; i < m_ComponentPools.size(); ++i) {
+                            m_ComponentPools[i]->remove(id);
+                            m_Systems[i]->onComponentRemovedFromEntity(entity);
+                        }
                     }
                     m_DestroyedEntities.clear();
-                }
-                for (auto& component_pool : m_ComponentPools) {
-                    component_pool->reserve(150);
+                    for (auto& component_pool : m_ComponentPools) {
+                        component_pool->reserve(150);
+                    }
                 }
             }
 
@@ -178,13 +161,8 @@ namespace Engine::priv {
                 return std::move(res);
             }
             void removeEntity(ENTITY entity) {
-                m_DestroyedEntities.push_back(entity);
+                m_DestroyedEntities.emplace_back(entity);
             }
-            /*
-            Entity getEntity(unsigned int entityID) const {
-                return m_EntityPool.getEntity(entityID); 
-            }
-            */
             template<class T, typename... ARGS> void addComponent(ENTITY entity, ARGS&&... args) {
                 auto type_slot = ECSRegistry::type_slot_fast<T>();
                 auto& cPool    = *static_cast<ECSComponentPool<ENTITY, T>*>(m_ComponentPools[type_slot]);
@@ -192,18 +170,6 @@ namespace Engine::priv {
                 {
                     std::lock_guard lock(m_Mutex);
                     res = cPool.addComponent(entity, std::forward<ARGS>(args)...);
-                }
-                if (res) {
-                    onComponentAddedToEntity(res, entity, type_slot);
-                }
-            }
-            template<class T, typename... ARGS> void addComponent(const EntityDataRequest& request, ENTITY entity, ARGS&&... args) {
-                auto type_slot = ECSRegistry::type_slot_fast<T>();
-                auto& cPool    = *static_cast<ECSComponentPool<ENTITY, T>*>(m_ComponentPools[type_slot]);
-                T* res         = nullptr;
-                {
-                    std::lock_guard lock(m_Mutex);
-                    res = cPool.addComponent(request, entity, std::forward<ARGS>(args)...);
                 }
                 if (res) {
                     onComponentAddedToEntity(res, entity, type_slot);
@@ -223,23 +189,15 @@ namespace Engine::priv {
                 return ret_val;
             }
 #pragma region 1 component get
-            template<class T> inline T* getComponent(ENTITY& entity) const {
+            template<class T> inline T* getComponent(ENTITY entity) const {
                 using CPoolType = ECSComponentPool<ENTITY, T>;
                 return static_cast<CPoolType*>(m_ComponentPools[ECSRegistry::type_slot_fast<T>()])->getComponent(entity);
-            }
-            template<class T> inline T* getComponent(const EntityDataRequest& dataRequest) const {
-                using CPoolType = ECSComponentPool<ENTITY, T>;
-                return static_cast<CPoolType*>(m_ComponentPools[ECSRegistry::type_slot_fast<T>()])->getComponent(dataRequest);
             }
 #pragma endregion
 
 #pragma region variadic component get
-            template<class... Types> inline std::tuple<Types*...> getComponents(ENTITY& entity) const {
-                EntityDataRequest dataRequest(entity);
-                return ECS::getComponents<Types...>(dataRequest);
-            }
-            template<class... Types> inline std::tuple<Types*...> getComponents(const EntityDataRequest& dataRequest) const {
-                return std::make_tuple(getComponent<Types>(dataRequest)... );
+            template<class... Types> inline std::tuple<Types*...> getComponents(ENTITY entity) const {
+                return std::make_tuple(getComponent<Types>(entity)...);
             }
 #pragma endregion
     };
